+++ src/js/3p-filters.js
@@ -74,7 +74,9 @@ const renderNodeStats = (used, total) => {
 };
 
 const i18nGroupName = name => {
-    return i18n$('3pGroup' + name.charAt(0).toUpperCase() + name.slice(1));
+    const groupname = i18n$('3pGroup' + name.charAt(0).toUpperCase() + name.slice(1));
+    if ( groupname !== '' ) { return groupname; }
+    return `${name.charAt(0).toLocaleUpperCase}${name.slice(1)}`;
 };
 
 /******************************************************************************/
@@ -223,6 +225,8 @@ const renderFilterLists = ( ) => {
             'privacy',
             'malware',
             'multipurpose',
+            'cookies',
+            'social',
             'annoyances',
             'regions',
             'custom'
@@ -235,9 +239,6 @@ const renderFilterLists = ( ) => {
         }
         for ( const [ listkey, listDetails ] of Object.entries(response.available) ) {
             let groupKey = listDetails.group;
-            if ( groupKey === 'social' ) {
-                groupKey = 'annoyances';
-            }
             const groupDetails = listTree[groupKey];
             if ( listDetails.parent !== undefined ) {
                 let lists = groupDetails.lists;
@@ -253,6 +254,14 @@ const renderFilterLists = ( ) => {
                 groupDetails.lists[listkey] = listDetails;
             }
         }
+        // https://github.com/uBlockOrigin/uBlock-issues/issues/3154#issuecomment-1975413427
+        //   Remove empty sections
+        for ( const groupkey of groupKeys ) {
+            const lists = listTree[groupkey].lists;
+            if ( Object.keys(lists).length !== 0 ) { continue; }
+            delete listTree[groupkey];
+        }
+
         const listEntries = createListEntries('root', listTree);
         qs$('#lists .listEntries').replaceWith(listEntries);
 
@@ -530,6 +539,34 @@ dom.on('#lists', 'click', 'span.cache', onPurgeClicked);
 /******************************************************************************/
 
 const selectFilterLists = async ( ) => {
+    // External filter lists to import
+    // Find stock list matching entries in lists to import
+    const toImport = (( ) => {
+        const textarea = qs$('#lists .listEntry[data-role="import"].expanded textarea');
+        if ( textarea === null ) { return ''; }
+        const lists = listsetDetails.available;
+        const lines = textarea.value.split(/\s+\n|\s+/);
+        const after = [];
+        for ( const line of lines ) {
+            if ( /^https?:\/\//.test(line) === false ) { continue; }
+            for ( const [ listkey, list ] of Object.entries(lists) ) {
+                if ( list.content !== 'filters' ) { continue; }
+                if ( list.contentURL === undefined ) { continue; }
+                if ( list.contentURL.includes(line) === false ) {
+                    after.push(line);
+                    continue;
+                }
+                const listEntry = qs$(`[data-key="${list.group}"] [data-key="${listkey}"]`);
+                if ( listEntry === null ) { break; }
+                toggleFilterList(listEntry, true);
+                break;
+            }
+        }
+        dom.cl.remove(textarea.closest('.expandable'), 'expanded');
+        textarea.value = '';
+        return after.join('\n');
+    })();
+
     // Cosmetic filtering switch
     let checked = qs$('#parseCosmeticFilters').checked;
     vAPI.messaging.send('dashboard', {
@@ -567,14 +604,6 @@ const selectFilterLists = async ( ) => {
         } else {
             listDetails.off = true;
         }
-    }
-
-    // External filter lists to import
-    const textarea = qs$('#lists .listEntry[data-role="import"].expanded textarea');
-    const toImport = textarea !== null && textarea.value.trim() || '';
-    if ( textarea !== null ) {
-        dom.cl.remove(textarea.closest('.expandable'), 'expanded');
-        textarea.value = '';
     }
 
     hashFromListsetDetails();
@@ -679,8 +708,6 @@ dom.on('.searchbar input', 'input', searchFilterLists);
 
 const expandedListSet = new Set([
     'uBlock filters',
-    'AdGuard – Annoyances',
-    'EasyList – Annoyances',
 ]);
 
 const listIsExpanded = which => {